requirements.txt
assemblyline_v4_service
assemblyline_service_client
vmray-rest-api==6.0.0
